Moved errors to their own file
